ClienteDAO, RegLivrosDAO, LivrosDAO e EstanteDAO implementados

+++ src/main/java/com/systemhaus/demo/dao/memory/EstanteDAO.java
@@ -38,7 +38,7 @@ public class EstanteDAO implements EstanteRepository {
 	}
 	
 	@Override
-	public int getCountOfEstantes() {
+	public long getCountOfEstantes() {
 		return biblioteca.getEstantes().size();
 	}
